Use package-dependant function ID
Now represented as u64 (see #53):

63           31           0
---------------------------
| package ID |  local ID  |
---------------------------

+++ src/mir/mir.rs
@@ -22,7 +22,7 @@ pub struct Function {
 }
 
 pub type LocalId = usize; // For now NameId are used as LocalId
-pub type FunctionId = usize;
+pub type FunctionId = u64;
 
 pub struct Local {
     pub id: LocalId,
